mcp_server/requirements.txt
fastmcp==2.11.3
uvicorn[standard]==0.35.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1